LLM_Bridge/requirements.txt
accelerate==1.9.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.10.0
asgiref==3.9.1
attrs==25.3.0
backoff==2.2.1
bcrypt==4.3.0
beautifulsoup4==4.13.4
bitsandbytes==0.46.1
blinker @ file:///home/conda/feedstock_root/build_artifacts/blinker_1731096409132/work
Brotli @ file:///home/conda/feedstock_root/build_artifacts/brotli-split_1749229842835/work
bs4==0.0.2
build==1.3.0
cachetools==5.5.2
certifi @ file:///home/conda/feedstock_root/build_artifacts/certifi_1752493286596/work/certifi
cffi @ file:///home/conda/feedstock_root/build_artifacts/cffi_1725560564262/work
charset-normalizer @ file:///home/conda/feedstock_root/build_artifacts/charset-normalizer_1746214863626/work
chroma-hnswlib==0.7.3
chromadb==0.5.3
click @ file:///home/conda/feedstock_root/build_artifacts/click_1747811314515/work
coloredlogs==15.0.1
cryptography==45.0.6
dataclasses-json==0.6.7
datasets==4.0.0
dill==0.3.8
distro==1.9.0
dnspython==2.7.0
duckduckgo_search==8.1.1
durationpy==0.10
email_validator==2.2.0
emoji==2.14.1
faiss-cpu==1.12.0
fastapi==0.111.0
fastapi-cli==0.0.8
filelock @ file:///home/conda/feedstock_root/build_artifacts/filelock_1741969488311/work
filetype==1.2.0
Flask @ file:///home/conda/feedstock_root/build_artifacts/flask_1747203985407/work
flatbuffers==25.2.10
frozenlist==1.7.0
fsspec==2025.3.0
gmpy2 @ file:///home/conda/feedstock_root/build_artifacts/gmpy2_1745509358281/work
google-auth==2.40.3
googleapis-common-protos==1.70.0
greenlet==3.2.4
grpcio==1.74.0
h11==0.16.0
h2 @ file:///home/conda/feedstock_root/build_artifacts/h2_1738578511449/work
hf-xet==1.1.5
hpack @ file:///home/conda/feedstock_root/build_artifacts/hpack_1737618293087/work
html5lib==1.1
httpcore==1.0.9
httptools==0.6.4
httpx==0.27.0
huggingface-hub==0.34.1
humanfriendly==10.0
hyperframe @ file:///home/conda/feedstock_root/build_artifacts/hyperframe_1737618333194/work
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1733211830134/work
importlib_metadata @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_importlib-metadata_1747934053/work
importlib_resources==6.5.2
itsdangerous @ file:///home/conda/feedstock_root/build_artifacts/itsdangerous_1733308265247/work
Jinja2 @ file:///home/conda/feedstock_root/build_artifacts/jinja2_1741263328855/work
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
kubernetes==33.1.0
langchain==0.2.6
langchain-community==0.2.6
langchain-core==0.2.43
langchain-text-splitters==0.2.2
langdetect==1.0.9
langsmith==0.1.147
lxml==6.0.0
markdown-it-py==4.0.0
MarkupSafe @ file:///home/conda/feedstock_root/build_artifacts/markupsafe_1733219680183/work
marshmallow==3.26.1
mdurl==0.1.2
mmh3==5.2.0
mpmath @ file:///home/conda/feedstock_root/build_artifacts/mpmath_1733302684489/work
multidict==6.6.3
multiprocess==0.70.16
mypy_extensions==1.1.0
networkx @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_networkx_1749078300/work
nltk==3.9.1
numpy==1.26.4
oauthlib==3.3.1
olefile==0.47
ollama==0.2.1
onnxruntime==1.22.1
opentelemetry-api==1.36.0
opentelemetry-exporter-otlp-proto-common==1.36.0
opentelemetry-exporter-otlp-proto-grpc==1.36.0
opentelemetry-instrumentation==0.57b0
opentelemetry-instrumentation-asgi==0.57b0
opentelemetry-instrumentation-fastapi==0.57b0
opentelemetry-proto==1.36.0
opentelemetry-sdk==1.36.0
opentelemetry-semantic-conventions==0.57b0
opentelemetry-util-http==0.57b0
optimum==1.26.1
orjson==3.11.1
overrides==7.7.0
packaging==24.2
pandas==2.3.1
pgvector==0.4.1
pillow @ file:///home/conda/feedstock_root/build_artifacts/pillow_1751481992986/work
posthog==6.5.0
primp==0.15.0
propcache==0.3.2
protobuf==6.31.1
psutil==7.0.0
psycopg==3.2.9
psycopg-binary==3.2.9
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_pycparser_1733195786/work
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
PyMuPDF==1.26.3
pypdf==4.2.0
PyPika==0.48.9
pyproject_hooks==1.2.0
PySocks @ file:///home/conda/feedstock_root/build_artifacts/pysocks_1733217236728/work
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-frontmatter==1.1.0
python-iso639==2025.2.18
python-magic==0.4.27
python-multipart==0.0.20
python-oxmsg==0.0.2
pytz==2025.2
PyYAML @ file:///home/conda/feedstock_root/build_artifacts/pyyaml_1737454647378/work
RapidFuzz==3.13.0
regex==2024.11.6
requests==2.32.3
requests-file==2.1.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.1.0
rich-toolkit==0.15.0
rsa==4.9.1
safetensors==0.5.3
sentencepiece==0.2.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.43
starlette==0.37.2
sympy==1.13.1
tenacity==8.5.0
tldextract==5.3.0
tokenizers==0.21.2
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
tqdm==4.67.1
transformers==4.54.0
triton==3.1.0
typer==0.16.0
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_typing_extensions_1751643513/work
tzdata==2025.2
ujson==5.10.0
unstructured==0.18.13
unstructured-client==0.42.3
urllib3 @ file:///home/conda/feedstock_root/build_artifacts/urllib3_1750271362675/work
uvicorn==0.30.1
uvloop==0.21.0
watchfiles==1.1.0
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
Werkzeug @ file:///home/conda/feedstock_root/build_artifacts/werkzeug_1733160440960/work
wrapt==1.17.2
xxhash==3.5.0
yarl==1.20.1
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1749421620841/work
zstandard==0.23.0
